+++ src/Traits/WithColumnSelect.php
@@ -31,17 +31,19 @@ trait WithColumnSelect
             ->filter(function ($column) {
                 return $column->isVisible() && $column->isSelectable() && $column->isSelected();
             })
-            ->map(fn ($column) => $column->getHash())
+            ->map(fn ($column) => $column->getSlug())
             ->values()
             ->toArray();
 
         // Set to either the default set or what is stored in the session
-        $this->selectedColumns = session()->get($this->getColumnSelectSessionKey(), $columns);
+        $this->selectedColumns = (isset($this->{$this->tableName}['columns']) && count($this->{$this->tableName}['columns']) > 0) ?
+            $this->{$this->tableName}['columns'] :
+            session()->get($this->getColumnSelectSessionKey(), $columns);
 
         // Check to see if there are any excluded that are already stored in the enabled and remove them
         foreach ($this->getColumns() as $column) {
-            if (! $column->isSelectable() && ! in_array($column->getHash(), $this->selectedColumns, true)) {
-                $this->selectedColumns[] = $column->getHash();
+            if (! $column->isSelectable() && ! in_array($column->getSlug(), $this->selectedColumns, true)) {
+                $this->selectedColumns[] = $column->getSlug();
                 session([$this->getColumnSelectSessionKey() => $this->selectedColumns]);
             }
         }
@@ -49,6 +51,7 @@ trait WithColumnSelect
 
     public function updatedSelectedColumns(): void
     {
-        session([$this->getColumnSelectSessionKey() => $this->selectedColumns]);
+        $this->{$this->tableName}['columns'] = $this->selectedColumns;
+        session([$this->getColumnSelectSessionKey() => $this->{$this->tableName}['columns']]);
     }
 }